pyproject.toml
[project]
name = "model-srs-comparision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "backoff>=2.2.1",
    "chromadb>=1.4.1",
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "ipywidgets>=8.1.8",
    "langchain-text-splitters>=1.1.0",
    "loguru>=0.7.3",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.53.0",
    "streamlit-mermaid>=0.3.0",
    "tabulate>=0.9.0",
    "uvicorn>=0.40.0",
    "weaviate-client>=4.19.2",
]
